Change TIMEOUT type to u64

+++ swap/src/api.rs
@@ -90,8 +90,8 @@ impl SwapLock {
     /// # Notes
     /// The 50ms polling interval is considered negligible overhead compared to the typical time required to suspend ongoing swap processes.
     pub async fn send_suspend_signal(&self) -> Result<(), Error> {
-        const TIMEOUT: i32 = 10_000;
-        const INTERVAL: i32 = 50;
+        const TIMEOUT: u32 = 10_000;
+        const INTERVAL: u64 = 50;
 
         let _ = self.suspension_trigger.send(())?;
